update cu_gstreamer and cu_dynthreshold to use optional dependency
