+++ components/tasks/cu_dynthreshold/Cargo.toml
@@ -12,7 +12,10 @@ homepage.workspace = true
 repository.workspace = true
 
 [dependencies]
-gstreamer = "0.23.5"
+gstreamer = { version = "0.23.5", optional = true }
 cu29 = { workspace = true }
 cu-sensor-payloads = { path = "../../payloads/cu_sensor_payloads", version = "0.7.0" }
-cu-gstreamer = { path = "../../sources/cu_gstreamer", version = "0.7.0" }
+cu-gstreamer = { path = "../../sources/cu_gstreamer", version = "0.7.0", optional = true }
+
+[features]
+gst = [ "dep:gstreamer", "dep:cu-gstreamer" ]
